size_t for gc globals (#1393)
Cherry-picked from HaxeFoundation/hxcpp e4dc00151e682940224becd988adffdec7cf4708.
Applied by hand rather than via cherry-pick: our copies of GC.h and GcCommon.cpp
use CRLF against upstream LF, so the automatic merge conflicted on whole files.

- sgTargetFreeSpacePercentage becomes size_t (sgMinimumWorkingMemory and
  sgMinimumFreeSpace were already size_t here)
- HXCPP_MINIMUM_WORKING_MEMORY / HXCPP_MINIMUM_FREE_SPACE / HXCPP_TARGET_FREE_SPACE
  are parsed with std::sscanf("%zu") and now raise a CriticalError on an
  unparseable value instead of silently ignoring it
- drop the now-redundant (size_t) casts in the working-memory calculations
- remove the dead sgObject_root global and its InitAlloc setup

Deviation from upstream: kept the (size_t) cast on the
baseMem * profileCollectSummary.spaceFactor expression under
HXCPP_GC_DYNAMIC_SIZE. spaceFactor is a double, so upstream dropping that cast
leaves std::max with no deducible template argument and does not compile in
that configuration.

Not build-verified: this machine cannot build the engine.

+++ include/hx/GC.h
@@ -165,7 +165,7 @@ extern size_t sgMinimumWorkingMemory;
 extern size_t sgMinimumFreeSpace;
 
 // Also ensure that the free memory is larger than this amount of used memory
-extern int sgTargetFreeSpacePercentage;
+extern size_t sgTargetFreeSpacePercentage;
 
 
 extern HXCPP_EXTERN_CLASS_ATTRIBUTES int gByteMarkID;
